refactor(swiftui): split OntologyBrowser into focused files (#1703)
